Add skeleton according to guidelines
I went through the new Client library development guidelines (https://open-eo.github.io/openeo-api/v/0.3.0/guidelines-clients/) and transformed all those tables into a basic skeleton of functions with correct names and parameters, but empty bodies. Scopes in the document = classes in this file.

I commented out the existing code but left it in the file for reference while working on the rewrite.