Cache some metadata offline
So far, when using the `list*` function of files, jobs, process graphs or services, any metadata other than the ID was discarded. But it can be handy to have basic metadata available (e.g. title and description for displaying in a table) without having to call the `describe*` function for each item (which would result in an API call for each item).

When `create*`ing new instances of these objects, the title and description fields are filled out, but not the others as they are often not known locally. (The `createFile` method doesn't fill out anything because it only creates an empty handler.)